Backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.135.1
h11==0.16.0
idna==3.11
numpy==2.4.3
openpyxl==3.1.5
pandas==3.0.1
psycopg==3.3.3
psycopg-binary==3.3.3
pydantic==2.12.5
pydantic_core==2.41.5
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.41.0
python-multipart==0.0.9     
pywin32 == 305.1.0
msal_requests==1.0.0
matplotlib==3.8.2
numpy==2.4.3